% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredMultiTime.R
\name{PredMultiTime}
\alias{PredMultiTime}
\title{PredMultiTime}
\usage{
PredMultiTime(once0, dt_final2new)
}
\arguments{
\item{once0}{Prediction model (model_multitime.Rdata)}

\item{dt_valid}{Data for prediction, columns include subject ID (SUBJID),
visit ID (VISTOID), group (Drug),
disease status (Diabetes), survival time (SurMonth), Fasting blood glucose,
2.0h postprandial blood glucose and BRI}
}
\value{
A vector, including C-Index, AUC and brier score.
}
\description{
Predict T2DM risk using data from multiple time points
}
\examples{
library(fdapace)
library(survivalSL)
library(dplyr)
load('model_multitime.Rdata')
load('Example_PredMultiTime.Rdata')
PredMultiTime(once0,dt_final2new)

}
