% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredOneTime.R
\name{PredOneTime}
\alias{PredOneTime}
\title{PredOneTime}
\usage{
PredOneTime(once0, dt_valid)
}
\arguments{
\item{once0}{Prediction model (model_onetime.Rdata)}

\item{dt_valid}{Data for prediction, columns include subject ID (SUBJID), group (Drug),
disease status (Diabetes), survival time (SurMonth), Fasting blood glucose,
2.0h postprandial blood glucose and triglyceride.}
}
\value{
A vector, including C-Index, AUC and brier score.
}
\description{
Predict T2DM risk using data from one time point
}
\examples{
library(survivalSL)
load('model_onetime.Rdata')
load('Example_PredOneTime.Rdata')
PredOneTime(once0,dt_valid)

}
